requirements.txt
Chameleon==2.13.post1
gpxpy==1.1.2
LatLon==1.0.2
Mako==0.9.0
-e git+git@github.com:cincodenada/mapthing.git@9bd170f1ab4e651bbce8d5613e2a5033f8fbe32a#egg=MapThing-dev
MarkupSafe==1.0
Paste==2.0.3
PasteDeploy==1.5.2
Pillow==4.1.1
Pygments==2.2.0
SQLAlchemy==1.2.5
WebOb==1.8.0rc1
argparse==1.2.1
gpxpy==1.2.0
hupper==1.1
olefile==0.44
plaster==1.0
plaster-pastedeploy==0.5
pyproj==1.9.5.1
pyramid==1.9.1
pyramid-chameleon==0.1
pyramid-debugtoolbar==1.0.9
pyramid-mako==0.3.1
pyramid-tm==0.7
pyshp==1.2.3
python-dateutil==2.6.0
repoze.lru==0.7
six==1.11.0
transaction==2.2.1
translationstring==1.3
venusian==1.1.0
waitress==0.8.7
wsgiref==0.1.2
zope.deprecation==4.3.0
zope.interface==4.4.3
zope.sqlalchemy==1.0
